VIPER RT: Change text parameter max length from `u16` to `u32`.
This should give us more room for the max length to not cause problems, but also fits better to Protobuf, which does not have a `u16` equivalent.

+++ opendut-viper/viper-rt/src/runtime/types/compile/parameters/mod.rs
@@ -182,7 +182,7 @@ pub enum ParameterDescriptor {
         name: ParameterName,
         info: ParameterInfo,
         default: Option<String>,
-        max: u16,
+        max: u32,
     }
 }
